feat(marketplace-period-filter): Add period filter to marketplace screen
Add a period filter to the marketplace screen in order to allow users to filter collections based on a specific time period. The period filter will be implemented using a dropdown menu with options for different time periods, such as "Last 10 minutes", "Last 1 hour", "Last 6 hours", etc. The selected time period will be used to retrieve collections that were created within that period.

Changes made in this commit:
- Added a new optional field "period_in_minutes" to the "GetCollectionsRequest" message in the "marketplace.proto" file.
- Added a new state variable "timePeriod" to the "UIStates" interface in the "marketplaceFilters.ts" file.
- Added a new array "marketplacePeriodItems" to define the available time period options in the "marketplaceFilters.ts" file.
- Added a new action "setTimePeriod" to update the selected time period in the