feat: remove a single wallet (#969)
* feat(wallets): add removeWalletById action

Add UI

* fix(profile): can't click dropdown items
adjust positionStyle in DropdownWithListItem (feat/wallet-single-delete)

* fix(wallet-single-delete): Navigate away if no wallet exists

Add logic to select default wallet when no wallet is selected

* feat(Profile): Add ability to delete a wallet (feat/wallet-single-delete)

* feat(wallet-single-delete): Add useEffect hook to TokenScreen.tsx

Add useEffect hook to TokenScreen.tsx to handle the scenario where the
selected wallet is not available. If there is no selected wallet and
there are wallets available, set the selected wallet to the first one.
If there are no wallets available, navigate to the "NativeWallet"
screen. This change ensures that the TokenScreen component behaves
appropriately in these scenarios.

* Update packages/screens/Mini/Profile/AccountDetailsScreen.tsx

Co-authored-by: n0izn0iz <n0izn0iz@users.noreply.github.com>

* feat(wallet-single-delete): Update TokenScreen types

---------

Co-authored-by: n0izn0iz <n0izn0iz@users.noreply.github.com>